backend/requirements.txt
django==3.2.3
psycopg2

dj-database-url==0.5.0
gunicorn==19.9.0
djangorestframework
flake8
pytest==4.4.0
pytest-django==3.4.8
pytest-factoryboy==2.0.2
requests
bs4
lxml
django-cors-headers
selenium
celery==4.4.1
django-redis
django-celery-beat
django_celery_results
django-json-widget
